@@ -9,16 +9,15 @@ use Symfony\Bundle\FrameworkBundle\Controller\AbstractController;
 use Symfony\Component\HttpFoundation\JsonResponse;
 use Symfony\Component\HttpFoundation\Request;
 use Symfony\Component\HttpFoundation\Response;
+use Symfony\Component\Serializer\SerializerInterface;
 
 class UpdatePresenceController extends AbstractController
 {
-    private $em;
-    private $smsService;
-    public function __construct(EntityManagerInterface $em, SmsService $smsService)
-    {
-        $this->em = $em;
-        $this->smsService = $smsService;
-    }
+    public function __construct(
+        private readonly EntityManagerInterface $em,
+        private readonly SmsService $smsService,
+        private readonly SerializerInterface $serializer
+    ) {}
 
     public function __invoke(Request $request): Response
     {
@@ -31,18 +30,20 @@ class UpdatePresenceController extends AbstractController
                 'status' => 'error',
                 'message' => 'Invalid params'
             ], Response::HTTP_BAD_REQUEST);
-        } else {
-            $personne->setPresent($content['present']);
-            $this->em->persist($personne);
-            $this->em->flush();
-
-//            if ($personne->getSmsSended() !== true) {
-            if ($content['present'] === true) {
-                $this->smsService->sendSms($personne);
-            }
-//            }
-
-            return new JsonResponse($personne);
         }
+
+        $personne->setPresent($content['present']);
+        $this->em->persist($personne);
+        $this->em->flush();
+
+        $forceSms = $content['withSms'] === true && $content['present'] === true;
+
+        if ($personne->getSmsSended() !== true || $forceSms) {
+            $this->smsService->sendSms($personne);
+        }
+
+        $serializedPersonne = $this->serializer->serialize($personne, 'json');
+
+        return new JsonResponse(json_decode($serializedPersonne));
     }
 }
