@@ -1,5 +1,5 @@
 /* eslint-disable no-nested-ternary */
-import React, { useEffect, useState } from 'react'
+import React, { useCallback, useEffect, useState } from 'react'
 import { useGetMany } from '../../hooks'
 import Table from './Table'
 
@@ -8,12 +8,29 @@ function TableProvider(props) {
   const [planElement, setPlanElement] = useState(null)
   const [planSize, setPlanSize] = useState(null)
   const [loading, setLoading] = useState(true)
-  const { items: allPersonnes, load: loadPersonnes } = useGetMany(
+  const { items: fetchedPersonnes, load: loadPersonnes } = useGetMany(
     `personnes?exists[table]=false`
   )
 
+  // État local pour gérer la liste des personnes non affectées
+  const [allPersonnes, setAllPersonnes] = useState([])
+
   useEffect(() => {
     loadPersonnes()
+  }, [])
+
+  // Synchroniser les données fetchées vers l'état local avec fullname
+  useEffect(() => {
+    const personnesWithFullname = fetchedPersonnes.map((item) => ({
+      ...item,
+      fullname: `${item.prenom} ${item.nom}`
+    }))
+    setAllPersonnes(personnesWithFullname)
+  }, [fetchedPersonnes])
+
+  // Callback pour retirer une personne de la liste après ajout réussi
+  const handlePersonneAdded = useCallback((personneId) => {
+    setAllPersonnes((prev) => prev.filter((p) => p.id !== personneId))
   }, [])
 
   useEffect(() => {
@@ -24,13 +41,6 @@ function TableProvider(props) {
       }, 100)
     }
   }, [props])
-
-  useEffect(() => {
-    allPersonnes.forEach((item) => {
-      // eslint-disable-next-line no-param-reassign
-      item.fullname = `${item.prenom} ${item.nom}`
-    })
-  }, [allPersonnes])
 
   useEffect(() => {
     if (
@@ -57,6 +67,7 @@ function TableProvider(props) {
               planSize={planSize}
               planRef={plan}
               allPersonnes={allPersonnes}
+              onPersonneAdded={handlePersonneAdded}
             />
           ))
         ) : (
